test: Split FieldMethodsTest into own file
